+++ src/Backoffice/index.js
@@ -1,5 +1,1 @@
-import TextLink from '../TextLink';
-import Button from './Button';
-import Tooltip from './Tooltip';
-
-export { TextLink, Button, Tooltip };
+export { default as Button } from './Button';