+++ src/GoogleAddressInputWithLabel/GoogleAddressInputWithLabel.driver.js
@@ -1,13 +1,13 @@
 import inputAreaWithLabelCompositeDriverFactory from '../Composite/InputAreaWithLabelComposite/InputAreaWithLabelComposite.driver';
 
-const GoogleAddressInputWithLabelDriverFactory = ({ element, wrapper }) => {
+const GoogleAddressInputWithLabelDriverFactory = ({ element }) => {
   const input = element.childNodes[1];
 
   return {
-    ...inputAreaWithLabelCompositeDriverFactory({ element, wrapper }),
+    ...inputAreaWithLabelCompositeDriverFactory({ element }),
     getInput: () => input,
     hasInput: () => input.childNodes[0].tagName.toLowerCase() === 'input',
   };
 };
 
-export default GoogleAddressInputWithLabelDriverFactory;
+export default googleAddressInputWithLabelDriverFactory;